requirements.txt
faster-whisper
pyannote.audio
openai
pydub
